refactor : selectFace → detect 로 패키지명 변경과 DetectFaceFragment로 파일명 변경 및 디자인 수정
