+++ presentation/src/main/res/layout/fragment_detect_face.xml
@@ -26,20 +26,21 @@
             android:id="@+id/bt_group_select"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:paddingHorizontal="8dp"
-            android:paddingVertical="4dp"
+            android:layout_marginEnd="16dp"
             android:fontFamily="@font/pretendard_regular"
             android:insetTop="0dp"
             android:insetBottom="0dp"
             android:letterSpacing="0"
             android:minWidth="0dp"
             android:minHeight="0dp"
+            android:paddingHorizontal="8dp"
+            android:paddingVertical="4dp"
             android:text="@string/all_select_text"
             android:textColor="@color/gray6"
             android:textSize="12sp"
             app:backgroundTint="@color/gray1"
             app:cornerRadius="20dp"
-            app:layout_constraintEnd_toEndOf="@+id/iv_initialization"
+            app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintTop_toTopOf="@+id/tv_detected_face_count"
             app:rippleColor="@android:color/transparent"
             app:strokeColor="@color/gray2"
@@ -77,7 +78,7 @@
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:src="@drawable/ic_initialization_25_25"
-            app:layout_constraintEnd_toEndOf="@+id/sd_detect_face"
+            app:layout_constraintEnd_toEndOf="@+id/bt_group_select"
             app:layout_constraintTop_toTopOf="@+id/tv_detect_face"
             tools:ignore="ContentDescription" />
 
@@ -85,7 +86,6 @@
             android:id="@+id/sd_detect_face"
             android:layout_width="0dp"
             android:layout_height="wrap_content"
-            android:layout_marginHorizontal="16dp"
             android:layout_marginBottom="26dp"
             android:stepSize="10"
             android:valueFrom="10"
