+++ presentation/src/main/res/values/dimens.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources xmlns:tools="http://schemas.android.com/tools">
     <!-- Material Slider -->
-    <dimen name="mtrl_slider_track_side_padding" tools:override="true">2.5dp</dimen>
+    <dimen name="mtrl_slider_track_side_padding" tools:override="true">16dp</dimen>
     <dimen name="mtrl_slider_widget_height" tools:override="true">16dp</dimen>
     <dimen name="mtrl_slider_track_height" tools:override="true">5dp</dimen>
     <dimen name="mtrl_slider_thumb_radius" tools:override="true">8dp</dimen>
